add sequicity for camrest and multiwoz, but 1) input file of sequicity not correct; 2) interface copy from Sequicity.py
