+++ tatk/e2e/sequicity/camrest/configs/camrest.json
@@ -1,0 +1,39 @@
+{
+  "tsdf_init":{
+    "beam_len_bonus": 0.5,
+    "prev_z_method": "separate",
+    "vocab_size": 800,
+    "embedding_size": 50,
+    "hidden_size": 50,
+    "split": [3,1,1],
+    "lr": 0.003,
+    "lr_decay": 0.5,
+    "vocab_path": "camrest/data/vocab.pkl",
+    "data": "camrest/data/CamRest676.json",
+    "entity": "camrest/data/CamRestOTGY.json",
+    "db": "camrest/data/CamRestDB.json",
+    "glove_path": "camrest/data/glove.6B.50d.txt",
+    "batch_size": 32,
+    "z_length": 8,
+    "degree_size": 5,
+    "layer_num": 1,
+    "dropout_rate": 0.5,
+    "epoch_num": 100,
+    "rl_epoch_num": 2,
+    "cuda": true,
+    "spv_proportion": 100,
+    "max_ts": 40,
+    "early_stop_count": 3,
+    "new_vocab": true,
+    "model_path": "camrest/output/sequicity_camrest.pkl",
+    "result_path": "camrest/output/camrest.csv",
+    "teacher_force": 100,
+    "beam_search": false,
+    "beam_size": 10,
+    "sampling": false,
+    "unfrz_attn_epoch": 0,
+    "skip_unsup": false,
+    "truncated": false,
+    "pretrain": false
+  }
+}
