--- tatk/e2e/sequicity/camrest/data/CamRest676/README.json
@@ -1,28 +1,0 @@
-#####################################################
-#####################################################
-#  Copyright Cambridge Dialogue Systems Group, 2016 #
-#####################################################
-#####################################################
-CamRest676 dataset contains the following three json files:
-1. CamRest676.json: the woz dialogue dataset, which contains the conversion from users and wizards, as well as a set of coarse labels for each user turn.
-2. CamRestDB.json: the Cambridge restaurant database file, containing restaurants in the Cambridge UK area and a set of attributes.
-3. The ontology file, specific all the values the three informable slots can take.
-
-
-The dataset was used in the following papers, if you use the dataset please cite them:
-
-@article{wenN2N16,
-       Author = {Wen, Tsung-Hsien and Vandyke, David and Mrk{\v{s}}i\'c, Nikola and Ga{\v{s}}i\'c, Milica and M. Rojas-Barahona, Lina and Su, Pei-Hao and Ultes, Stefan and Young, Steve},
-       title={A Network-based End-to-End Trainable Task-oriented Dialogue System},
-       journal={arXiv preprint: 1604.04562},
-       year={2016},
-       month={April}
-}
-
-@article{wencond16,
-       Author = {Wen, Tsung-Hsien and Ga{\v{s}}i\'c, Milica and Mrk{\v{s}}i\'c, Nikola and M. Rojas-Barahona, Lina and Su, Pei-Hao and Ultes, Stefan and Vandyke, David and Young, Steve},
-       title={Conditional Generation and Snapshot Learning in Neural Dialogue Systems},
-       journal={arXiv preprint: 1606.03352},
-       year={2016},
-       month={June}
-}
