+++ tatk/e2e/sequicity/multiwoz/configs/multiwoz.json
@@ -1,0 +1,47 @@
+{
+  "tsdf_init":{
+    "beam_len_bonus": 0.5,
+    "prev_z_method": "separate",
+    "vocab_size": 4000,
+    "embedding_size": 50,
+    "hidden_size": 50,
+    "split": [3,1,1],
+    "lr": 0.003,
+    "lr_decay": 0.5,
+    "vocab_path": "multiwoz/data/vocab-multiwoz.pkl",
+    "train": "multiwoz/data/train.json",
+    "dev": "multiwoz/data/val.json",
+    "test": "multiwoz/data/test.json",
+    "entity": "multiwoz/data/entities.json",
+    "db": [
+      "../../../data/multiwoz/db/attraction_db.json",
+      "../../../data/multiwoz/db/hotel_db.json",
+      "../../../data/multiwoz/db/restaurant_db.json",
+      "../../../data/multiwoz/db/hospital_db.json",
+      "../../../data/multiwoz/db/train_db.json"
+    ],
+    "glove_path": "multiwoz/data/glove.6B.50d.txt",
+    "batch_size": 32,
+    "z_length": 8,
+    "degree_size": 5,
+    "layer_num": 1,
+    "dropout_rate": 0.5,
+    "epoch_num": 100,
+    "rl_epoch_num": 2,
+    "cuda": true,
+    "spv_proportion": 100,
+    "max_ts": 40,
+    "early_stop_count": 3,
+    "new_vocab": true,
+    "model_path": "multiwoz/output/sequicity_multiwoz.pkl",
+    "result_path": "multiwoz/output/multiwoz.csv",
+    "teacher_force": 100,
+    "beam_search": false,
+    "beam_size": 10,
+    "sampling": false,
+    "unfrz_attn_epoch": 0,
+    "skip_unsup": false,
+    "truncated": false,
+    "pretrain": false
+  }
+}
